// src/app/_core/models/cri/intervention-report-vo.ts
import { EntityBase } from '.././entity-base';
import { ReferenceDataVO } from '../../models';
import { PostalAddressVO, UserVO, ContactMethodVO, CustomerVO } from '../models';
import { DocumentVO } from '../documentVO';
import { InterventionDetailVO } from './intervention-detail-vo';
import { InterventionHardwareVO } from './intervention-hardware-vo';


export interface InterventionReportVO extends EntityBase {

		id: number;
		technicians: UserVO[];
		interventionHardware: InterventionHardwareVO[];
		interventionDetail: InterventionDetailVO[];
		postalAdress: PostalAddressVO;
		phoneNumber: ContactMethodVO;
		otherActors: string[];
		otherParticipent: UserVO[];
		atCustomerStartTime: Date;
		interventionStartTime: Date;
		interventionEndTime: Date;
		atCustomerEndTime: Date;
		atCustomerStartTimeStr: string;
		interventionStartTimeStr: string;
		interventionEndTimeStr: string;
		atCustomerEndTimeStr: string;
		pseStartTime: Date;
		pseEndTime: Date;
		pseStartTimeStr: string;
		pseEndTimeStr: string;
		reportOnServiceBill: string;
		isPaymentDone: boolean;
		isBillingAccorderToQuotation: boolean;
		billingComment: string;
		isHno: boolean;
		isHno2 : boolean;
		isOnCallDuty: boolean;
		isPublicHoliday: boolean;
		customerId: number;
		customerLastNameFirstName: string;
		refInitiator: ReferenceDataVO;
		manualInitiatorIntervention: string;
		taskId: number;
		coachName: string;
		offerCustomerLabel: string;
		customerStatus: number;
		accountManagerName: string;
		isRequestClose: boolean;
		document: DocumentVO;
		customer: CustomerVO;
		requestId: number;
		isSatisfied: boolean;
		unsatisfactionComment: string;
		isCoachPresent: boolean;
		additionalDocuments: DocumentVO[];
		isOnTwoDays : boolean;
		interventionStartTimeDay2: Date;
		interventionEndTimeDay2: Date;
		atCustomerStartTimeDay2: Date;
		atCustomerEndTimeDay2: Date;
		interventionStartTimeDay2Str: string;
		interventionEndTimeDay2Str: string;
		atCustomerStartTimeDay2Str: string;
		atCustomerEndTimeDay2Str: string;
		isOnCallDuty2 : boolean;
		isPublicHoliday2: boolean;
		createdBy : UserVO;
		 contactName : string;
		contactFirstName: string;
		isCoachContact: boolean;
		isMembreContact: boolean;
	  }